feat: enhance MSEController with improved error handling and logging; migrate exception handling to TypeScript
